% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{check_for_local_pg}
\alias{check_for_local_pg}
\title{Check if a performance graph exists locally}
\usage{
check_for_local_pg(workout_id)
}
\arguments{
\item{workout_id}{the ID to check.}
}
\value{
lgl
}
\description{
Check if a performance graph exists locally
}
