% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{data_dir}
\alias{data_dir}
\title{Return a local storage path}
\usage{
data_dir(dir_type = c("pg", "wl", "uid"))
}
\arguments{
\item{dir_type}{"pg" for performance graphs, "wl" for workout list, and uid
for "user_id"}
}
\value{
character with file path
}
\description{
Performance graphs, workout lists, and your local user ID are stored
locally. This function returns the directory where these things are stored.
}
\examples{
data_dir("pg")
}
