% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_performance_graph}
\alias{get_performance_graph}
\title{Get performance graph for a workout ID}
\usage{
get_performance_graph(workout_id, save = TRUE)
}
\arguments{
\item{workout_id}{The workout ID to get}

\item{save}{should this be saved locally?}
}
\value{
performance graph result
}
\description{
Get performance graph for a workout ID
}
