% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{pelo_get_workout_list}
\alias{pelo_get_workout_list}
\title{Get a list of all workouts from Peloton}
\usage{
pelo_get_workout_list(user_id = pelo_user_id(), n_to_get = 10000)
}
\arguments{
\item{user_id}{The user_id}

\item{n_to_get}{Max number of workouts to get}
}
\value{
tibble with list of all workouts
}
\description{
Get a list of all workouts from Peloton
}
