% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{pelo_init_local_data_store}
\alias{pelo_init_local_data_store}
\title{Initialize the local data store directory}
\usage{
pelo_init_local_data_store(save_to = pelo_local_dir())
}
\arguments{
\item{save_to}{directory to create local storage folders}
}
\description{
This package will save peloton data locally to the desired location.
By default this is ~/peloton_data.
}
\details{
This initializes that directory.
}
