% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{pelo_login}
\alias{pelo_login}
\title{Log in to your peloton account in order to access privileged data, e.g. workout details}
\usage{
pelo_login()
}
\value{
a response
}
\description{
Log in to your peloton account in order to access privileged data, e.g. workout details
}
