% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{pelo_performance_graph}
\alias{pelo_performance_graph}
\title{Load the performance graph for a workout ID}
\usage{
pelo_performance_graph(workout_id, save = TRUE)
}
\arguments{
\item{workout_id}{the workout ID to get the PG for}

\item{save}{whether to save locally}
}
\value{
a performance graph response
}
\description{
Load the performance graph for a workout ID
}
