% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{pelo_update_performance_graphs}
\alias{pelo_update_performance_graphs}
\title{Update performance graphs}
\usage{
pelo_update_performance_graphs(sleep = 2, force_refresh = FALSE)
}
\arguments{
\item{sleep}{how long to sleep between API calls}

\item{force_refresh}{refresh workout list?}
}
\description{
Update performance graphs
}
