% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{pelo_user_id}
\alias{pelo_user_id}
\title{Get Peloton user_id}
\usage{
pelo_user_id()
}
\value{
the Peloton User ID
}
\description{
Note that the user_id is not the same as your login credentials. This will
check for a locally stored user ID, and if there's not one, it will fetch
the user_id by signing into Peloton.
}
