% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{pelo_workout_list}
\alias{pelo_workout_list}
\title{Get the list of all workouts}
\usage{
pelo_workout_list(
  force_refresh = FALSE,
  user_id = pelo_user_id(),
  n_to_get = 10000
)
}
\arguments{
\item{force_refresh}{force remote fetch rather than local}

\item{user_id}{user ID}

\item{n_to_get}{how many workouts to get}
}
\value{
a tibble with workouts
}
\description{
Get the list of all workouts
}
